lonestar output in	mem-latency 			mem-transpose 			omp-latency 			omp-transpose
all this output is roughly the same as on screms after the 2.67 --->
3.33 CPUGHZ conversion.

bandwidths are a bit lower on lonestar and latency is roughly the same

the low copy bandwidth in mem-bw is almost certainly because of icpc
ver 11 vs ver 12. Ver 11 appears to either use a worse version of
intel_fast_memcopy or is generating worse code elsewhere.